Display flash messages when a meeting is updated
Because this was a 'res.render', rather than a 'res.redirect',
we had to include not only a req.flash('success', msg) statement,
but also the key/value pair 'flashes: req.flash()'. Otherwise,
the flash message wouldn't display until there was a redirect or
a hard refresh of the page.