Convert Geolocation to google.maps.LatLng Class
For reasons unknown, map.setCenter/map.panTo is not currently
repositioning the map to the user's geolocation. This, in spite
of the fact that the object seems to be created successfully.
We'll need to resolve this to move forward.

Google's API required that we do some special formatting on the
geolocation numbers. Specifically, we had to ensure that there
were no more than 2 numeric characters each for latitude and
longitude. Additionally, we had to account for the fact that there
are sometimes negative numbers (adding a "-" character to the mix).
Regex helped resolve this.