Set Map Center As User Geolocation (If Available)
The `map.panTo(lat,lng)` function was not working when called in
the `makeMap()` function itself. By moving it into the
`loadPlaces()` function - where meeting markers are placed and
bounds are set - the command took its desired effect.

I admit that I'm not 100% sure of what the issue here was. I believe
it has something to do with the fact that `map` is being filled and
positioned in `loadPlaces()`. (But it should be able to be
repositioned later, should it not?)