Set Map to Regional Zoom with Geolocation
Now, when the map is centered on the user's geolocation, the map also
zooms to a regional level. This is hardcoded at 9.

For the future, it would be good to have the zoom level depend on
what markers were (or weren't) placed in the immediate vicinity of
the geolocation center. Tighter for higher meeting density, wider
for areas with few or no meetings.