Update Routes and loadPages API to Allow Search
This commit renames our 2 existing API routes and creates a new one.
With these routes in place, the loadPages API is updated to take the
slug and criteria parameters.

Following these changes, the next step is to create a searchMeetings
module in meetingController. This will allow us to display search
results on the map, according to parameters (slug and criteria) passed
to updateMap().