Add searchMeetings Controller & Return Query ID
This commit adds a controller function that returns a key/value pair.
This pair will provide information needed for searching the database
and returning the appropriate entries for a specific map.

Currently, res.json() is returning '{ slug : "query-string" }', with
slug being a fixed key. Our next task is to make this key dynamic,
being populated by whatever param happens to used in the URL.