Use API Queries Properly to Return Meetings
This update has the API returning meetings based on queries, rather
than relying on params to serve as the 'key' in the key/value pair.

The API can currently return results based on a single criteria (e.g.,
'yearlymeeting'). The next step is to make the API capable of returning
multi-criteria searches.