Handle for Both Single and Multi-Criteria Searches
From a button click on the main menu, we can cause the main map to
display the results of searches with multiple queries. The map will
display all meetings that meet all conditions.

The next step is probably to build out an interface that allows the
user to select and search for meetings according to their criteria.