Add Dropdowns for Yearly Meeting, State, Worship Style, and Branch
This is the first step in creating an interface that allows the user
to update the map according to selection criteria. In this commit, we
add a few dropdown boxes whose options are populated by whatever values
are available in the database. However, there are a couple problems
here:

* The queries we're running are very inefficient - one for each
dropdown. There must be a way to do these four operations in just one,
but I'm not yet sure how Mongo allows for this.

* The options are being returned in non-alphabetical order. That doesn't
work for me, but Mongo's .sort() doesn't seem to return what I need.

Next steps here are to resolve these issues, as able. More importantly,
we need to build out this user interface and get it to dynamically
update the main map.