Return Sorted Options
In order to return the dropdown options in alphabetical order, I needed
to assign my variables in a different way. Instead of assigning each
set of options from the database call itself, I assigned them in the
callback. I'm not sure why the former way doesn't work, but the latter
seems to work just fine.

This commit also includes some updates to the dropdown forms that
corrects labels and gives each option the appropriate value.

The question of a more efficient solution for this database call is
still an open one. For the time being, I'm satisfied with my present
solution.